Refactor & Bugfixes
