+++ src/i18n/locales/en.json
@@ -91,13 +91,14 @@
     "retry": "Do you like to generate a new package?",
     "rightsInfo": "All rights by tekom",
     "rulesChecked": "rules checked for validation",
-    "showOnlyViolations": "show only first {0} violations",
     "settingsRestored": "Restored default settings",
+    "showOnlyViolations": "show only first {0} violations",
     "startNewPackage": "New package",
     "thankYou": "Thank you!",
     "thankYouExplainer": "Your iiRDS-Package is now ready for saving.",
     "valid": "Valid",
-    "violationsDetected": "violations detected for file",
+    "violationInstancesDetected": "occurences in total",
+    "violationsDetected": "One violation detected for file | {n} violation detected for file | {n} violations detected for file",
     "welcomeSubtitle": "With the free iiRDS Validation Tool you can validate iiRDS packages in just a few steps",
     "welcomeTitle": "Welcome to the iiRDS Validation Tool!"
   },
